backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
gunicorn
flask-cors
redis